Move dimension-related CSS vars to their own file
