+++ src/components/Workspace/SideBarItem.vue
@@ -1,6 +1,6 @@
 <template>
-  <div class="side-bar-item" @click="click()">
-    <Icon v-if="icon" :icon="icon" size="small" :url="url" />
+  <div class="side-bar-item">
+    <Icon v-if="icon" :icon="icon" size="small" />
     <p v-else>{{ title }}</p>
   </div>
 </template>
@@ -8,6 +8,7 @@
 <script>
 
 import Icon from '@/components/LinkItems/ItemIcon.vue';
+import Defaults from '@/utils/defaults';
 
 export default {
   name: 'SideBarItem',
@@ -15,11 +16,20 @@ export default {
   props: {
     icon: String,
     title: String,
-    url: String,
-    click: Function,
+  },
+  mounted() {
+    this.initiateFontAwesome();
   },
   components: {
     Icon,
+  },
+  methods: {
+    initiateFontAwesome() {
+      const fontAwesomeScript = document.createElement('script');
+      const faKey = this.config.appConfig.fontAwesomeKey || Defaults.fontAwesomeKey;
+      fontAwesomeScript.setAttribute('src', `https://kit.fontawesome.com/${faKey}.js`);
+      document.head.appendChild(fontAwesomeScript);
+    },
   },
 };
 </script>
